bq-demo/requirements.txt
cryptography==2.1.4
google-cloud-bigquery==1.24.0
google-oauth==1.0.1
google-auth==1.16.0
